backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
openai==1.12.0
python-socketio==5.11.0
python-dotenv==1.0.0
eventlet==0.35.1
